including quick start example of how to configure APIM with OAuth 2.0 with Entra ID App Roles
Including how to configure APIM with OAuth 2.0 with Entra ID using App Roles for Authentication and Authorization for an custom application (e.g. SPA). 
This is very usefull and common scenario that I work a lot in customers when you have legacy APIs, APIs with no OAuth 2.0 / OIDC or even if you need to centralized management of authentication and authentication in one single pane of glass.
Very common in API Gateways architectures